mother-duck-rag/back/pyproject.toml
[project]
name = "back-rag"
version = "0.1.0"
description = "RAG backend: ingestão + API para vault Obsidian (DuckDB)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.0.0",
    "fastapi>=0.115.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "sentence-transformers>=2.2.0",
    "tabulate>=0.9.0",
    "tqdm>=4.65.0",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
